+++ crates/bevy_mod_chroma_api/src/lib.rs
@@ -8,7 +8,7 @@ use bevy::{
 };
 use plugin::ApplyEffectRequest;
 use reqwest::Url;
-use serde::Serialize;
+use serde::{Deserialize, Serialize};
 
 pub mod api;
 pub mod bgr_color;
@@ -51,7 +51,7 @@ impl<'w, 's> Chroma<'w, 's> {
     }
 }
 
-#[derive(Debug)]
+#[derive(Debug, Clone, Copy, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
 pub struct EffectHandle {
     entity: Entity,
 }
@@ -75,7 +75,7 @@ impl ChromaRunner {
     }
 }
 
-#[derive(Resource, Debug, Clone)]
+#[derive(Resource, Debug, Clone, Serialize, Deserialize, Eq, PartialEq, Ord, PartialOrd, Hash)]
 pub struct ChromaRunnerInitializationSettings {
     init_url: &'static str,
     init_request: InitRequest,
@@ -94,17 +94,28 @@ impl ChromaRunnerInitializationSettings {
     }
 }
 
-#[derive(Debug, Serialize, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize, Eq, PartialEq, Ord, PartialOrd, Hash)]
 pub struct InitRequest {
     pub title: &'static str,
     pub description: &'static str,
     pub author: Author,
-    pub device_supported: Vec<&'static str>,
+    pub device_supported: Vec<SupportedDevice>,
     pub category: &'static str,
 }
 
-#[derive(Debug, Serialize, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize, Eq, PartialEq, Ord, PartialOrd, Hash)]
 pub struct Author {
     pub name: &'static str,
     pub contact: &'static str,
+}
+
+#[derive(Debug, Copy, Clone, Serialize, Deserialize, Eq, PartialEq, Ord, PartialOrd, Hash)]
+#[serde(rename_all = "lowercase")]
+pub enum SupportedDevice {
+    Keyboard,
+    Mouse,
+    Mousepad,
+    Headset,
+    Keypad,
+    ChromaLink,
 }