+++ crates/bevy_mod_chroma_api/src/plugin.rs
@@ -7,7 +7,7 @@ use bevy::{
     utils::Instant,
 };
 use bevy_mod_chroma_request_lib::{
-    HttpRequestError, HttpRequestHandle, HttpRequestPlugin, HttpRequests,
+    HttpRequestError, HttpRequestHandle, HttpRequestPlugin, HttpRequestSet, HttpRequests,
 };
 use serde::Serialize;
 
@@ -30,23 +30,33 @@ impl Plugin for ChromaPlugin {
                 ),
             )
             .add_system(
-                system_create_pending_effects.run_if(
-                    resource_exists::<ChromaRunner>().and_then(in_state(RunnerState::Running)),
-                ),
+                system_create_pending_effects
+                    .in_base_set(HttpRequestSet::BeforeExecuteRequests)
+                    .run_if(
+                        resource_exists::<ChromaRunner>().and_then(in_state(RunnerState::Running)),
+                    ),
             )
             .add_system(
-                system_gather_create_effect_results.run_if(
-                    resource_exists::<ChromaRunner>().and_then(in_state(RunnerState::Running)),
-                ),
+                system_gather_create_effect_results
+                    .in_base_set(HttpRequestSet::AfterGatherResponses)
+                    .run_if(
+                        resource_exists::<ChromaRunner>().and_then(in_state(RunnerState::Running)),
+                    ),
             )
             .add_system(
-                system_apply_effects.run_if(
-                    resource_exists::<ChromaRunner>().and_then(in_state(RunnerState::Running)),
-                ),
+                system_apply_effects
+                    .in_base_set(HttpRequestSet::BeforeExecuteRequests)
+                    .run_if(
+                        resource_exists::<ChromaRunner>().and_then(in_state(RunnerState::Running)),
+                    ),
             )
-            .add_system(system_apply_effects_cleanup.run_if(
-                resource_exists::<ChromaRunner>().and_then(in_state(RunnerState::Running)),
-            ));
+            .add_system(
+                system_apply_effects_cleanup
+                    .in_base_set(HttpRequestSet::AfterGatherResponses)
+                    .run_if(
+                        resource_exists::<ChromaRunner>().and_then(in_state(RunnerState::Running)),
+                    ),
+            );
     }
 }
 
@@ -92,7 +102,7 @@ fn system_init(
     // SAFETY: init_request is always Some here as verified above
     if let Some(response) = requests.get_response(init_request.as_ref().unwrap()) {
         let session_info = response.as_ref()?.json::<SessionInfo>()?;
-        let root_url = if session_info.root_url.ends_with("/") {
+        let root_url = if session_info.root_url.ends_with('/') {
             session_info.root_url
         } else {
             session_info.root_url + "/"
@@ -111,31 +121,31 @@ fn system_init(
         info!("successfully opened chroma session to {}", root_url);
     }
 
-    return Ok(());
+    Ok(())
 }
 
 #[derive(Debug)]
 enum InitError {
-    RequestError(HttpRequestError),
-    ParseError(serde_json::Error),
-    UrlError(url::ParseError),
+    Request(HttpRequestError),
+    Parse(serde_json::Error),
+    Url(url::ParseError),
 }
 
 impl From<&HttpRequestError> for InitError {
     fn from(error: &HttpRequestError) -> Self {
-        Self::RequestError(error.clone())
+        Self::Request(error.clone())
     }
 }
 
 impl From<serde_json::Error> for InitError {
     fn from(error: serde_json::Error) -> Self {
-        Self::ParseError(error)
+        Self::Parse(error)
     }
 }
 
 impl From<url::ParseError> for InitError {
     fn from(error: url::ParseError) -> Self {
-        Self::UrlError(error)
+        Self::Url(error)
     }
 }
 
@@ -149,6 +159,7 @@ struct CreatedEffect {
     id: String,
 }
 
+#[allow(clippy::type_complexity)]
 fn system_create_pending_effects(
     mut commands: Commands,
     mut requests: HttpRequests,