+++ src/lib.rs
@@ -1,5 +1,11 @@
-pub use bevy_mod_chroma_api::{Author, ChromaRunnerInitializationSettings, InitRequest};
+pub use bevy_mod_chroma_api::{
+    Author, ChromaRunnerInitializationSettings, InitRequest, SupportedDevice,
+};
 
 pub use bevy_mod_chroma_api::{Chroma, ChromaPlugin, EffectHandle};
 
-pub use bevy_mod_chroma_api::api::{Effect, KeyboardEffect, MouseEffect};
+pub use bevy_mod_chroma_api::api::Effect;
+
+pub use bevy_mod_chroma_api::api::{
+    ChromaLinkEffect, HeadsetEffect, KeyboardEffect, KeypadEffect, MouseEffect, MousepadEffect,
+};